Extract theme page; add assets & theme route
Move the large theme/color-system UI out of BrandShowcase into a new ThemeShowcase page at /brand/theme and simplify the main /brand page to link to it. Add a dedicated theme route page (brand/theme/page.tsx) and register it in the sitemap. Update BrandShowcase to streamline the brand kit UI: add altFile support for dual downloads (PNG + SVG), show file-type labels, and adjust download layout. Add SVG wordmark/mark assets and update the brand-kit zip, plus CSS tweaks for the theme link and download row. Also update the /brand page metadata/description to reflect the new wording.
